refactor(compiler): reorganise files
